+++ math/src/main/kotlin/ktx/math/ranges.kt
@@ -125,7 +125,7 @@ fun ClosedRange<Float>.randomTriangular() = MathUtils.randomTriangular(start, en
  * is treated as exclusive as it is not practical to keep it inclusive. Results are undefined for an empty range, and
  * there is no error checking.
  */
-fun ClosedRange<Float>.randomTriangular(normalizedMode: Float) =
+fun ClosedRange<Float>.randomTriangular(normalizedMode: Float): Float =
   MathUtils.randomTriangular(start, endInclusive,
     normalizedMode * (endInclusive - start) + start
   )
@@ -137,7 +137,7 @@ fun ClosedRange<Float>.randomTriangular(normalizedMode: Float) =
  * [ClosedRange.endInclusive].
  * @return The interpolated value.
  */
-fun ClosedRange<Float>.lerp(progress: Float) =
+fun ClosedRange<Float>.lerp(progress: Float): Float =
   progress * (endInclusive - start) + start
 
 /**
@@ -148,5 +148,5 @@ fun ClosedRange<Float>.lerp(progress: Float) =
  * @param interpolation The function to interpolate with.
  * @return The interpolated value.
  */
-fun ClosedRange<Float>.interpolate(progress: Float, interpolation: Interpolation) =
+fun ClosedRange<Float>.interpolate(progress: Float, interpolation: Interpolation): Float =
   interpolation.apply(progress) * (endInclusive - start) + start